kafka_python_gcp/pyproject.toml
[project]
name = "kafka-python-gcp"
version = "0.1.0"
description = "Kafka integration with Python for ML pipelines and GCP services"
requires-python = ">=3.9"

dependencies = [
    # Kafka dependencies
    "kafka-python>=2.2.15",
    # Data processing
    "pandas>=2.0.3",
    "pyarrow>=12.0.1",
    "numpy>=1.21.0",
    "pandas-gbq>=0.26.1",
    # GCP dependencies
    "google-cloud-bigquery>=3.11.4",
    "google-auth>=2.23.3",
    "google-cloud-secret-manager>=2.16.4",
    "google-cloud-storage>=2.10.0",
    # Kubeflow Pipeline dependencies
    "kfp>=2.0.0",
    # Utility dependencies
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "pytest>=7.0.0"
]

[tool.setuptools]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]